v1.0.3 - 5

+++ org.emdepub/src/main/java/org/emdepub/ai_md/engine/AiMdEngine.java
@@ -2,12 +2,17 @@ package org.emdepub.ai_md.engine;
 
 import java.util.Arrays;
 
+import org.emdepub.activator.L;
 import org.emdepub.ai_md.call.AiMdCall;
+import org.emdepub.ai_md.engine.model.AiMdReferenceHeader;
 import org.emdepub.ai_md.parser.ext.references.AiMdReferenceBlock;
 import org.emdepub.ai_md.parser.ext.references.AiMdReferencesExtension;
 import org.emdepub.common.utils.CU;
+import org.emdepub.toml.editor.engine.TomlResSupport;
 import org.jetbrains.annotations.NotNull;
 
+import com.fasterxml.jackson.dataformat.toml.TomlMapper;
+import com.vladsch.flexmark.ast.Paragraph;
 import com.vladsch.flexmark.ast.Text;
 import com.vladsch.flexmark.ext.gfm.tasklist.TaskListExtension;
 import com.vladsch.flexmark.ext.tables.TablesExtension;
@@ -22,21 +27,20 @@ import com.vladsch.flexmark.util.data.DataHolder;
 import com.vladsch.flexmark.util.data.MutableDataHolder;
 import com.vladsch.flexmark.util.data.MutableDataSet;
 
-import lombok.AccessLevel;
-import lombok.experimental.FieldDefaults;
-
-@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class AiMdEngine {
 
-	static MutableDataSet parserOptions = new MutableDataSet();
+	
+	private static final int editorLine[] = new int[1];
+	
+	private static final MutableDataSet parserOptions = new MutableDataSet();
 	static {
 		parserOptions.set(Parser.EXTENSIONS, Arrays.asList(TablesExtension.create(), TaskListExtension.create(),
 				AiMdReferencesExtension.create(), AiMdReferenceBlockReplacingExtension.create()));	
 	}
-	static Parser parser = Parser.builder(parserOptions).build();
-	static Formatter formatter = Formatter.builder(parserOptions).build();
+	private static final Parser parser = Parser.builder(parserOptions).build();
+	private static final Formatter formatter = Formatter.builder(parserOptions).build();
 
-    static class AiMdReferenceBlockReplacingPostProcessor extends NodePostProcessor {
+	private static class AiMdReferenceBlockReplacingPostProcessor extends NodePostProcessor {
 
         static class Factory extends NodePostProcessorFactory {
 
@@ -55,26 +59,36 @@ public class AiMdEngine {
 
         @Override
         public void process(@NotNull NodeTracker state, @NotNull Node node) {
+        	
+        	
         	if (node instanceof AiMdReferenceBlock) {
         		AiMdReferenceBlock aiMdReferenceBlock = (AiMdReferenceBlock) node;
-        		
-        		String aiSourceText = aiMdReferenceBlock.getTextBlockContent().getContents().toStringOrNull();
-        		
-        		
-        		
-        		String aiResultText = AiMdCall.callRephrase(aiSourceText);
-        		
-                Text text = new Text(aiResultText);
-                aiMdReferenceBlock.insertAfter(text);
-                state.nodeAdded(text);
+        		if ((aiMdReferenceBlock.getHeaderEndLineNumber() < editorLine[0]) && (editorLine[0] < aiMdReferenceBlock.getEndLineNumber())) {
 
-                aiMdReferenceBlock.unlink();
-                state.nodeRemoved(aiMdReferenceBlock);
+        			String aiSourceHeaderText = aiMdReferenceBlock.getHeaderBlockContent().getContents().toStringOrNull();
+        			
+        			AiMdReferenceHeader aiMdReferenceHeader = CU.yamlDeserialize(aiSourceHeaderText, AiMdReferenceHeader.class);
+        			
+        			String aiSourceText = aiMdReferenceBlock.getTextBlockContent().getContents().toStringOrNull();
+            		String aiResultText = aiMdReferenceHeader.getId() + " -- " + AiMdCall.callRephrase(aiSourceText);
+                    
+            		Paragraph paragraph = new Paragraph();
+            		aiMdReferenceBlock.insertAfter(paragraph);
+            		state.nodeAdded(paragraph);
+            		
+            		Text text = new Text(aiResultText);
+            		paragraph.appendChild(text);
+            		state.nodeAdded(text);
+
+                    aiMdReferenceBlock.unlink();
+                    state.nodeRemoved(aiMdReferenceBlock);
+
+        		}
         	}
         }
     }
 
-    static class AiMdReferenceBlockReplacingExtension implements Parser.ParserExtension {
+	private static class AiMdReferenceBlockReplacingExtension implements Parser.ParserExtension {
 
         private AiMdReferenceBlockReplacingExtension() { }
 
@@ -91,8 +105,10 @@ public class AiMdEngine {
         }
     }
 	
-	public static void generateMdFromAiMd(String aiMdSource, String fileDestinationFilePath) {
+	public static void generateMdFromAiMd(String aiMdSource, String cursorPositionString, String fileDestinationFilePath) {
 
+		editorLine[0] = Integer.valueOf(CU.trimStrip(cursorPositionString.split(":")[0])) - 1;
+		
 		Node documentNodeReplacing = parser.parse(aiMdSource);
 		
 		CU.saveStringToFile(formatter.render(documentNodeReplacing), fileDestinationFilePath);
